URL post fix for "already in database"
Adding posts via URL will now do one of two things:
If post is already in the system, it will reset its status to RAW so it can be edited again. It will also update it's content and also the time gathered so it will be at the top of the "Recently added" filter in the client program.
If not already in the system, it will add it to the system.